requirements.txt
numpy
opencv-python
Pillow
tensorflow
streamlit